# packages/api-schema/openapi.yaml
openapi: 3.1.0
info: { title: VyNode Media API, version: 0.1.0, description: Versioned client contract. Breaking changes require a new /api/vN namespace. }
servers: [{ url: / }]
paths:
...
  /api/v1/playback/sessions:
    post: { operationId: createPlaybackSession, summary: "Selects the least-cost valid mode through authenticated HLS video transcode and returns an explainable quality/backend plan", security: [{ bearerAuth: [] }], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PlaybackStart' } } } }, responses: { '201': { description: Persistent playback session and structured decision }, '400': { $ref: '#/components/responses/BadRequest' }, '403': { description: Playback denied }, '503': { description: Transcode capacity reached } } }
  /api/v1/playback/sessions/{sessionId}:
    patch: { operationId: updatePlaybackSession, summary: Owner-only progress/state update, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/PlaybackSessionId' }], responses: { '204': { description: Progress persisted }, '403': { description: Not session owner } } }
    delete: { operationId: stopPlaybackSession, summary: Stops current user's playback session, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/PlaybackSessionId' }], responses: { '204': { description: Session stopped }, '403': { description: Not session owner } } }
  /api/v1/playback/sessions/{sessionId}/media:
    get: { operationId: deliverPlaybackMedia, summary: Authorized original bytes with Range or progressive generated fMP4 with validated start seconds, parameters: [{ $ref: '#/components/parameters/PlaybackSessionId' }, { name: Range, in: header, schema: { type: string } }, { name: start, in: query, schema: { type: number, minimum: 0, maximum: 604800 } }], responses: { '200': { description: Original or generated stream }, '206': { description: Direct Play byte range }, '403': { description: Invalid playback authorization }, '410': { description: Playback session ended }, '416': { description: "Invalid range or Range on generated media" }, '503': { description: Pipeline capacity reached or FFmpeg unavailable } } }
    head: { operationId: headPlaybackMedia, summary: Original media headers without body, parameters: [{ $ref: '#/components/parameters/PlaybackSessionId' }, { name: Range, in: header, schema: { type: string } }], responses: { '200': { description: Full media headers }, '206': { description: Range headers }, '416': { description: Invalid range } } }
  /api/v1/playback/sessions/{sessionId}/hls/{file}:
    get: { operationId: deliverPlaybackHlsResource, summary: "Session-authorized private HLS playlist, initialization fragment, or segment without exposing filesystem paths", parameters: [{ $ref: '#/components/parameters/PlaybackSessionId' }, { name: file, in: path, required: true, schema: { type: string, pattern: '^(master\\.m3u8|init\\.mp4|segment-[0-9]{6}\\.m4s)$' } }], responses: { '200': { description: Authorized HLS resource }, '403': { description: "Cross-user, expired, or invalid resource denied" }, '410': { description: Playback session ended }, '503': { description: Video transcode capacity reached } } }
  /api/v1/playback/{logicalType}/{logicalId}/versions:
    get: { operationId: listPlaybackVersions, summary: Safe technical versions without paths, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/LogicalType' }, { $ref: '#/components/parameters/LogicalId' }], responses: { '200': { description: Available versions }, '404': { $ref: '#/components/responses/NotFound' } } }
  /api/v1/playback/sessions/{sessionId}/subtitles/{trackId}:
    get: { operationId: deliverPlaybackSubtitle, summary: Session-authorized sanitized WebVTT, parameters: [{ $ref: '#/components/parameters/PlaybackSessionId' }, { name: trackId, in: path, required: true, schema: { type: string } }], responses: { '200': { description: WebVTT subtitle, content: { text/vtt: {} } }, '403': { description: Invalid playback authorization } } }
  /api/v1/playback/continue-watching:
    get: { operationId: continueWatching, summary: Current user's incomplete local progress, security: [{ bearerAuth: [] }], responses: { '200': { description: Partially watched movies and episodes }, '403': { description: Denied } } }
  /api/v1/playback/continue-watching/items/{logicalType}/{logicalId}:
    delete: { operationId: dismissContinueWatching, summary: Hides an item until that user plays it again without changing watched state, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/LogicalType' }, { $ref: '#/components/parameters/LogicalId' }], responses: { '204': { description: Item dismissed }, '403': { description: Denied } } }
  /api/v1/account/playback-preferences:
    get: { operationId: getPlaybackPreferences, summary: Current user's semantic audio subtitle autoplay and quality defaults, security: [{ bearerAuth: [] }], responses: { '200': { description: Playback preferences }, '403': { description: Denied } } }
    patch: { operationId: updatePlaybackPreferences, summary: Replaces validated preferences for the authenticated user only, security: [{ bearerAuth: [] }], responses: { '200': { description: Updated playback preferences }, '400': { $ref: '#/components/responses/BadRequest' } } }
  /api/v1/playback/{logicalType}/{logicalId}/markers:
    get: { operationId: listMediaMarkers, summary: Returns logical-timeline skip markers without paths, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/LogicalType' }, { $ref: '#/components/parameters/LogicalId' }], responses: { '200': { description: Manual marker ranges }, '403': { description: Denied } } }
  /api/v1/playback/{logicalType}/{logicalId}/start-over:
    post: { operationId: resetPlaybackProgress, summary: Resets resume position without conflating dismissal or watched state, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/LogicalType' }, { $ref: '#/components/parameters/LogicalId' }], responses: { '204': { description: Resume progress reset }, '403': { description: Denied } } }
  /api/v1/admin/media-markers:
    post: { operationId: createMediaMarker, summary: OWNER or ADMIN creates a validated manual logical-timeline marker and audit event, security: [{ bearerAuth: [] }], responses: { '201': { description: Marker created }, '400': { $ref: '#/components/responses/BadRequest' }, '403': { description: Denied } } }
  /api/v1/admin/media-markers/{markerId}:
    patch: { operationId: updateMediaMarker, summary: OWNER or ADMIN updates a manual marker and audits the change, security: [{ bearerAuth: [] }], parameters: [{ name: markerId, in: path, required: true, schema: { type: string } }], responses: { '200': { description: Marker updated }, '403': { description: Denied } } }
    delete: { operationId: deleteMediaMarker, summary: OWNER or ADMIN deletes a marker and audits the change, security: [{ bearerAuth: [] }], parameters: [{ name: markerId, in: path, required: true, schema: { type: string } }], responses: { '204': { description: Marker deleted }, '403': { description: Denied } } }
  /api/v1/playback/{logicalType}/{logicalId}/progress:
    get: { operationId: getPlaybackProgress, summary: Current user's private logical-media progress, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/LogicalType' }, { $ref: '#/components/parameters/LogicalId' }], responses: { '200': { description: Progress or empty initial state }, '403': { description: Denied } } }
  /api/v1/playback/{logicalType}/{logicalId}/watched:
    put: { operationId: setWatchedState, summary: Marks current user's item watched or unwatched, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/LogicalType' }, { $ref: '#/components/parameters/LogicalId' }], responses: { '204': { description: Watch state updated }, '400': { $ref: '#/components/responses/BadRequest' } } }
  /api/v1/admin/playback/sessions:
    get: { operationId: listActivePlaybackSessions, summary: Requires playback.sessions.view; no paths, security: [{ bearerAuth: [] }], responses: { '200': { description: Active streams }, '403': { description: Denied } } }
  /api/v1/admin/playback/capabilities:
    get: { operationId: getPlaybackCapabilities, summary: Safe FFmpeg capabilities and pipeline counts, security: [{ bearerAuth: [] }], responses: { '200': { description: FFmpeg version and relevant capabilities }, '403': { description: Denied } } }
  /api/v1/admin/playback/sessions/{sessionId}:
    delete: { operationId: terminatePlaybackSession, summary: Requires playback.sessions.manage and emits a security audit event, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/PlaybackSessionId' }], responses: { '204': { description: Session terminated }, '404': { $ref: '#/components/responses/NotFound' } } }
...
  /api/v1/admin/analytics/playback:
    get: { operationId: getAdminPlaybackAnalytics, summary: Requires observability.view; UTC range is bounded to 366 days, security: [{ bearerAuth: [] }], parameters: [{ name: from, in: query, schema: { type: string, format: date-time } }, { name: to, in: query, schema: { type: string, format: date-time } }, { name: userId, in: query, schema: { type: string } }], responses: { '200': { description: Server-wide playback analytics }, '400': { $ref: '#/components/responses/BadRequest' } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    BadRequest: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  parameters:
...
    PlaybackSessionId: { name: sessionId, in: path, required: true, schema: { type: string, format: uuid } }
    LogicalType: { name: logicalType, in: path, required: true, schema: { type: string, enum: [MOVIE, EPISODE] } }
    LogicalId: { name: logicalId, in: path, required: true, schema: { type: string, format: uuid } }
  schemas:
    PlaybackStart:
      type: object
      required: [logicalType, logicalId, resume, capabilities]
      properties:
        logicalType: { type: string, enum: [MOVIE, EPISODE] }
        logicalId: { type: string, format: uuid }
        requestedVersionId: { type: string, format: uuid }
        selectedAudioTrackId: { type: string }
        selectedSubtitleTrackId: { type: string }
        resume: { type: boolean }
        capabilities: { $ref: '#/components/schemas/ClientCapabilities' }
    ClientCapabilities:
      type: object
      required: [schemaVersion, clientName, platform, supportedContainers, supportedVideoCodecs, supportedAudioCodecs, directPlaySupport]
      properties:
        schemaVersion: { type: integer, enum: [1, 2] }
        clientName: { type: string }
        clientVersion: { type: string }
        platform: { type: string }
        platformVersion: { type: string }
        deviceModel: { type: string }
        supportedContainers: { type: array, maxItems: 20, items: { type: string } }
        supportedVideoCodecs: { type: array, maxItems: 20, items: { type: string } }
        supportedAudioCodecs: { type: array, maxItems: 30, items: { type: string } }
        subtitleFormats: { type: array, items: { type: string } }
        maximumVideoWidth: { type: integer, minimum: 0 }
        maximumVideoHeight: { type: integer, minimum: 0 }
        maximumAudioChannels: { type: integer, minimum: 0, maximum: 16 }
        fragmentedMp4Support: { type: boolean }
        hdrCapabilities: { type: array, items: { type: string } }
        directPlaySupport: { type: boolean }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties: { code: { type: string }, message: { type: string }, requestId: { type: string } }